Transitioned Regex extension to extension file
